+++ crates/objc2-encode/src/helper.rs
@@ -366,6 +366,9 @@ impl<E: EncodingType> Helper<'_, E> {
                 write!(f, "]")?;
             }
             Self::Container(kind, name, items) => {
+                if !verify_name(name) {
+                    panic!("{kind} name was not a valid identifier: {name:?}");
+                }
                 write!(f, "{}", kind.start())?;
                 write!(f, "{name}")?;
                 if let Some(level) = level.container_include_fields() {
@@ -453,18 +456,8 @@ impl Helper<'_> {
             Pointer(t) => Self::Indirection(IndirectionKind::Pointer, t),
             Atomic(t) => Self::Indirection(IndirectionKind::Atomic, t),
             Array(len, item) => Self::Array(*len, item),
-            Struct(name, fields) => {
-                if !verify_name(name) {
-                    panic!("Struct name was not a valid identifier");
-                }
-                Self::Container(ContainerKind::Struct, name, fields)
-            }
-            Union(name, members) => {
-                if !verify_name(name) {
-                    panic!("Union name was not a valid identifier");
-                }
-                Self::Container(ContainerKind::Union, name, members)
-            }
+            Struct(name, fields) => Self::Container(ContainerKind::Struct, name, fields),
+            Union(name, members) => Self::Container(ContainerKind::Union, name, members),
             None => Self::NoneInvalid,
         }
     }
@@ -504,18 +497,8 @@ impl<'a> Helper<'a, EncodingBox> {
             Pointer(t) => Self::Indirection(IndirectionKind::Pointer, t),
             Atomic(t) => Self::Indirection(IndirectionKind::Atomic, t),
             Array(len, item) => Self::Array(*len, item),
-            Struct(name, fields) => {
-                if !verify_name(name) {
-                    panic!("Struct name was not a valid identifier");
-                }
-                Self::Container(ContainerKind::Struct, name, fields)
-            }
-            Union(name, members) => {
-                if !verify_name(name) {
-                    panic!("Union name was not a valid identifier");
-                }
-                Self::Container(ContainerKind::Union, name, members)
-            }
+            Struct(name, fields) => Self::Container(ContainerKind::Struct, name, fields),
+            Union(name, members) => Self::Container(ContainerKind::Union, name, members),
             None => Self::NoneInvalid,
         }
     }