+++ crates/objc2-encode/CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Equivalence comparisons now consider `Encoding::String` as equivalent to
   `Encoding::Pointer(&Encoding::Char)`/`Encoding::Pointer(&Encoding::UChar)`.
 
+### Fixed
+* Relaxed assertions on `Encoding::Struct`/`Encoding::Union`'s name to allow
+  unicode identifiers.
+
 
 ## [4.1.0] - 2025-01-22
 [4.1.0]: https://github.com/madsmtm/objc2/compare/objc2-encode-4.0.3...objc2-encode-4.1.0